+++ src/core/component/directives/attrs/index.ts
@@ -39,6 +39,10 @@ import {
 import type { ComponentInterface } from 'core/component/interface';
 import type { DirectiveParams } from 'core/component/directives/attrs/interface';
 
+//#if runtime has dummyComponents
+import('core/component/directives/attrs/test/b-component-directives-attrs-dummy');
+//#endif
+
 export * from 'core/component/directives/attrs/const';
 export * from 'core/component/directives/attrs/interface';
 
@@ -225,8 +229,7 @@ ComponentEngine.directive('attrs', {
 				modifiers
 			};
 
-			const
-				cantIgnoreDir = value != null || decl.length !== 2;
+			const cantIgnoreDir = value != null || decl.length !== 2;
 
 			if (Object.isDictionary(dir)) {
 				if (Object.isFunction(dir.beforeCreate)) {
@@ -257,19 +260,14 @@ ComponentEngine.directive('attrs', {
 		}
 
 		function parseEventListener(attrName: string, attrVal: unknown) {
-			let
-				isDOMEvent = true,
-				event = attrName.slice(1).camelize(false);
+			let event = attrName.slice(1).camelize(false);
 
 			const
-				originalEvent = event,
-				eventChunks = event.split('.');
+				eventChunks = event.split('.'),
+				flags = Object.createDict<boolean>();
 
-			const
-				flags = Object.createDict<boolean>(),
-				isOnceEvent = flags.once;
-
-			eventChunks.forEach((chunk) => flags[chunk] = true);
+			// The first element is the event name; we need to slice only the part containing the event modifiers
+			eventChunks.slice(1).forEach((chunk) => flags[chunk] = true);
 			event = eventChunks[0];
 
 			if (flags.right && !event.startsWith('key')) {
@@ -281,12 +279,10 @@ ComponentEngine.directive('attrs', {
 
 			} else {
 				event = `on${event.capitalize()}`;
-				isDOMEvent = false;
 			}
 
 			if (flags.capture) {
 				event += 'Capture';
-				isDOMEvent = true;
 				delete flags.capture;
 			}
 
@@ -297,7 +293,6 @@ ComponentEngine.directive('attrs', {
 
 			if (flags.passive) {
 				event += 'Passive';
-				isDOMEvent = true;
 				delete flags.passive;
 			}
 
@@ -322,21 +317,7 @@ ComponentEngine.directive('attrs', {
 			const isSystemGetter = propGetterRgxp.test(event);
 			props[event] = attrVal;
 
-			if (componentCtx != null && !isDOMEvent && Object.isFunction(attrVal)) {
-				// Under the contract, all handlers must be in `$attrs`
-				componentCtx.unsafe.$attrs ??= {};
-				componentCtx.unsafe.$attrs[event] = attrVal;
-
-				if (!isSystemGetter) {
-					if (isOnceEvent) {
-						componentCtx.$on(originalEvent, attrVal);
-
-					} else {
-						componentCtx.$once(originalEvent, attrVal);
-					}
-				}
-
-			} else if (!isSystemGetter) {
+			if (!isSystemGetter) {
 				setVNodePatchFlags(vnode, 'events');
 
 				const dynamicProps = vnode.dynamicProps ?? [];