+++ src/core/component/directives/attrs/CHANGELOG.md
@@ -15,6 +15,19 @@ Changelog
 
 * Fixed a bug where adding `v-attrs` to components could cause them to re-render
 
+## v4.0.0-beta.?? (2024-??-??)
+
+#### :house: Internal
+
+* Removed dead code
+
+## v4.0.0-beta.110 (2024-07-17)
+
+#### :bug: Bug Fix
+
+* Fixed the bug where the event name was set as an event modifier in the `v-attrs` directive
+* Replaced the method calls to `componentCtx.$once` and `componentCtx.$on` with correct event handling based on the isOnceEvent flag
+
 ## v4.0.0-alpha.1 (2022-12-14)
 
 #### :rocket: New Feature